+++ src/main/java/com/briup/jz/service/impl/BaseRolePrivilegeServiceImpl.java
@@ -4,6 +4,7 @@ import com.briup.jz.bean.BaseRolePrivilege;
 import com.briup.jz.bean.BaseRolePrivilegeExample;
 import com.briup.jz.dao.BaseRolePrivilegeMapper;
 import com.briup.jz.service.IBaseRolePrivilegeService;
+import com.briup.jz.utils.CustomerException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -41,6 +42,9 @@ public class BaseRolePrivilegeServiceImpl implements IBaseRolePrivilegeService {
 
     @Override
     public void deleteById(Long id) {
+        if (baseRolePrivilegeMapper.selectByPrimaryKey(id) == null) {
+            throw new CustomerException("要删除的分类信息不存在");
+        }
         baseRolePrivilegeMapper.deleteByPrimaryKey(id);
     }
 }
